# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Node.js API with JWT Authentication",
    "description": "API for managing Users, Roles, Claims, and TestObjects with JWT-based Authentication"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "http://localhost:3001",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'"
      }
    },
    "schemas": {
...
      "Role": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "name": { "type": "string", "example": "admin" }
        }
      },
      "Claim": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "name": { "type": "string", "example": "CreateUser" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "example": "Unauthorized" },
          "status": { "type": "integer", "example": 401 }
        }
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/roles": {
      "get": {
        "tags": ["Roles"],
        "summary": "Get all roles",
        "description": "Retrieve a list of all roles (Admin only)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of roles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Role" }
                }
              }
            }
          },
          "400": { "description": "Bad request", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "500": { "description": "Server error", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } }
        }
      },
      "post": {
        "tags": ["Roles"],
        "summary": "Create a new role",
        "description": "Create a new role (Admin only)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Role created successfully" },
          "400": { "description": "Bad request", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "500": { "description": "Server error", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } }
        }
      }
    },
    "/roles/{id}": {
      "get": {
        "tags": ["Roles"],
        "summary": "Get a role by ID",
        "description": "Retrieve a role by its ID (Admin only)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Role ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Role retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Role" }
              }
            }
          },
          "400": { "description": "Bad request", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "404": { "description": "Role not found", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "500": { "description": "Server error", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } }
        }
      },
      "put": {
        "tags": ["Roles"],
        "summary": "Update role by ID",
        "description": "Update a role by its ID (Admin only)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Role ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Role updated successfully" },
          "400": { "description": "Bad request", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "404": { "description": "Role not found", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "500": { "description": "Server error", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } }
        }
      },
      "delete": {
        "tags": ["Roles"],
        "summary": "Delete a role by ID",
        "description": "Delete a role by its ID (Admin only)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Role ID"
          }
        ],
        "responses": {
          "200": { "description": "Role deleted successfully" },
          "400": { "description": "Bad request", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "404": { "description": "Role not found", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "500": { "description": "Server error", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } }
        }
      }
    },
    "/claims": {
      "get": {
        "tags": ["Claims"],
        "summary": "Get all claims",
        "description": "Retrieve a list of all claims (Admin only)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of claims",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Claim" }
                }
              }
            }
          },
          "400": { "description": "Bad request", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "500": { "description": "Server error", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } }
        }
      },
      "post": {
        "tags": ["Claims"],
        "summary": "Create a new claim",
        "description": "Create a new claim (Admin only)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Claim created successfully" },
          "400": { "description": "Bad request", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "500": { "description": "Server error", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } }
        }
      }
    },
    "/claims/{id}": {
      "get": {
        "tags": ["Claims"],
        "summary": "Get a claim by ID",
        "description": "Retrieve a claim by its ID (Admin only)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Claim ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Claim retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Claim" }
              }
            }
          },
          "400": { "description": "Bad request", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "404": { "description": "Claim not found", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "500": { "description": "Server error", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } }
        }
      },
      "put": {
        "tags": ["Claims"],
        "summary": "Update claim by ID",
        "description": "Update a claim by its ID (Admin only)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Claim ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Claim updated successfully" },
          "400": { "description": "Bad request", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "404": { "description": "Claim not found", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "500": { "description": "Server error", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } }
        }
      },
      "delete": {
        "tags": ["Claims"],
        "summary": "Delete a claim by ID",
        "description": "Delete a claim by its ID (Admin only)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Claim ID"
          }
        ],
        "responses": {
          "200": { "description": "Claim deleted successfully" },
          "400": { "description": "Bad request", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "404": { "description": "Claim not found", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } },
          "500": { "description": "Server error", "content": { "application/json": { "$ref": "#/components/schemas/Error" } } }
        }
      }
    },
...
  }
}